fix: expand mujoco node addition to URDF
Previously, the mujoco node was added to the URDF under conversion
without regard to whether there was one already. MuJoCo didn't like
that.

Now, the _one_ node that might be already in the URDF file and the _one_
that might be provided by the user override/update the default node. The
user-provided node takes precedence.